Add Sketchup files for extruder motor mount.
This is the bottom mount and filament path for the filament feeder on a
Nema17 motor. It is drawn in sketchup and exported as STL from there.
It seemed quicker to do it in SketchUp rather than OpenSCAD since the
Nema model was available, and drawing the feeder size was measure
directly. I needed this in a hurry, but should really be re-done in
OpenSCAD at some point :-)

The Nema17.skp file is from the 3D Warehouse.