Add optionis for newer E3D hotend mount to effector.
After damaging my E3D V5 (at least I think it is a V5) Chinese clone, I
had to modify the effector to accept a newer V6 clone. Due to the fan
duct on the newer hotend, I had to drop it down in the effector which
required a new mount clamp and also detachable Z-Probe mounting.

THe code is in a bit of a mess here and needs some clean up still.